integrated comp geo module
I replaced the calls to sympy in trajectory_tracker.py with my own comp
geo implementation with numpy. I also removed one of the calls to
tt.model.evaluate by replacing it with a shift from the previous call.
Now, the loop takes about 186271 microseconds for 500 iterations, which
is about 373 microseconds per iteration. The function tt.model.evaluate
takes on the order of 57 microseconds to evaluate, so the remaining 300
or so microseconds are probably taken up by my intersection method.