recorded a bunch of ball motion data
it's in the pkl and txt files. enjoy!
also, a note: sometimes the polling produces a bunch of errors several
times a second--i hypothesize that sometimes the server disconnects
without the polling being successfully stopped, but i'm not sure why
that it. i'll check it out later; it's not a huge problem since the
error doesn't happen until after the program has stopped doing the gist
of what it's doing.